requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg2-binary==2.9.9
SQLAlchemy==2.0.36
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.3.4
# alembic==1.13.3
asyncpg==0.29.0
pytest==8.3.2
httpx==0.27.0
databases==0.9.0
setuptools
cryptography==41.0.4